ga first imple, not tested, floats changed to doubles

+++ src/ttp/City.hpp
@@ -6,13 +6,13 @@ namespace ttp {
 
 struct City
 {
-	float getDistance(const City& other) const
+	double getDistance(const City& other) const
 	{
 		return sqrt((other.x - x) * (other.x - x) + (other.y - y) * (other.y - y));
 	}
 
 	uint32_t index;
-	float x;
-	float y;
+	double x;
+	double y;
 };
 } // namespace ttp
